+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 27/01-2019
+
+### Changed
+
+- Split Update-SteamApp into Install-SteamCMD.
+
+### Removed
+
+- Remove temp file with SteamApp IDs. Instead, the newest list is always downloaded.
+
 ## [1.0.3] - 27/01-2019
+
+### Added
 
 - Add -UseBasicParsing to Invoke-WebRequest in order to support Windows Server Core.